% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt2Convexhull}
\alias{dt2Convexhull}
\title{data.frame to data.table with column for convex hull geometry by ID}
\usage{
dt2Convexhull(d, pid, y = "y", x = "x", dt = "time", projection)
}
\arguments{
\item{d}{a data.frame including points which belong to clusters (tdbscan output), latitude, longitude and datetime}

\item{pid}{ID unique for points that belong to one cluster}

\item{y}{column name (character vector). The latitude.}

\item{x}{column name (character vector). The longitude.}

\item{dt}{column name (character vector).  datetime}

\item{projection}{character vector. if missing will default to  `"+proj=utm +zone=4 +datum=WGS84"` with a warning.}
}
\value{
a data.table with geometry column (convex hull polygons) and connected arrival and departure datetime
}
\description{
data.frame to data.table with column for convex hull geometry by ID
}
\examples{
require(data.table)
data(zbird)
z = tdbscan(zbird, eps = 12, minPts   = 5, maxLag = 5, borderPoints = TRUE )
z = z[, clustID := factor(clustID)]

o = data.frame(zbird) \%>\% data.table
o = merge(z, o, by.x = 'id', by.y = 'sp.ID')

o = rbindlist(list(copy(o[, tagID := 'bird1']), copy(o[, tagID := 'bird2'])), use.names = TRUE)
o[tagID == 'bird2', x := x + 5]
o[tagID == 'bird2', x := x + 5]
o[!is.na(clustID), ID := paste0(tagID, '_', clustID)]

d = dt2Convexhull(o, pid = 'ID', projection = '+proj=utm +zone=4 +datum=WGS84')

}
