% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt2Track}
\alias{dt2Track}
\title{data.frame to Track}
\usage{
dt2Track(d, y = "latitude", x = "longitude", dt = "time", projection)
}
\arguments{
\item{d}{a data.frame or an object inheriting from it}

\item{y}{column name (character vector). The latitude.}

\item{x}{column name (character vector). The longitude.}

\item{dt}{column name (character vector).  datetime}

\item{projection}{character vector. if missing will default to  `"+proj=longlat +ellps=WGS84"` with a warning.}
}
\value{
A  \link[trajectories]{Track}
}
\description{
data.frame to Track
}
\examples{

data(pesa56511)
x = data.frame(pesa56511) \%>\% data.table
o = dt2Track(x, projection = proj4string(pesa56511) )
a = tdbscan(pesa56511, eps =6600, minPts   = 8, maxLag = 6, borderPoints = TRUE )
b = tdbscan(o, eps =6600, minPts   = 8, maxLag = 6, borderPoints = TRUE )
identical(a, b)

}
