% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{pesa56511}
\alias{pesa56511}
\title{A dataset containing the track of a pectoral sandpiper male.}
\format{Track
\describe{
  \item{trajectories::Track}{with "connections","data","sp", "time", "endTime" slots}
  \item{quality}{in data slot, is the Argos location quality}
}}
\description{
A dataset containing the track of a pectoral sandpiper male.
}
\keyword{data}
