% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdbscan.R
\name{tdbscan}
\alias{tdbscan}
\title{tdbscan}
\usage{
tdbscan(track, eps, minPts = 5, borderPoints = FALSE, maxLag = 6,
  minTenure)
}
\arguments{
\item{track}{A \link[sp:is.projected]{projected}  \link[trajectories:Track]{Track}.}

\item{eps}{size of the epsilon neighborhood (see \code{\link[dbscan:dbscan]{dbscan::dbscan()}}  ).}

\item{minPts}{number of minimum points in the eps region (for core points).
Default is 5 points (see \code{\link[dbscan:dbscan]{dbscan::dbscan()}}  ).}

\item{borderPoints}{Logical; default to FALSE (see \code{\link[dbscan:dbscan]{dbscan::dbscan()}}  ).}

\item{maxLag}{maximum relative temporal lag (see notes). Default to 6.
When set to \code{maxLag>N} output is the same as for \link[dbscan:dbscan]{dbscan}.}

\item{minTenure}{minimum time difference, in hours, between the last and the first entry a cluster.
Clusters with  values smaller than minTenure are discarded.}
}
\description{
tdbscan
}
\note{
\strong{TODO}
}
\examples{
  require(tdbscan)
  require(sp)
  require(trajectories)
  require(ggplot2)
# Pectoral Sandpiper
  data(pesa56511)
  z = tdbscan(pesa56511, eps =6600, minPts   = 8, maxLag = 6, borderPoints = TRUE )
  
  o = cbind(coordinates(pesa56511), z)
  ggplot(o, aes(longitude/10000,latitude/10000,color = factor(clustID))) +
   geom_path(aes(color= NULL), col = 'grey', size = .5) + 
   geom_point( )

# toybird 
  data(toybird)   
  z = tdbscan(toybird, eps = 500, minPts   = 3, maxLag = 6)
  z[, .N, clustID]
  
  o = cbind(coordinates(toybird), z)[, clustID := factor(clustID)]
  
  ggplot(o, aes(coords.x1/100000,coords.x2/100000,color = clustID)) +
   geom_path(aes(color= NULL), col = 'grey', size = .5) + 
   geom_point( ) + geom_text((aes(label = id)), hjust = 1.5 ,check_overlap = TRUE)
  
  # set min tenure
 z = tdbscan(toybird, eps = 500, minPts = 3, maxLag = 6, minTenure = 5)
 z[, .N, clustID]


}
