% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{zbird}
\alias{zbird}
\title{A dataset containing the track of a z bird.}
\format{Track
\describe{
  \item{trajectories::Track}{with "connections","data","sp", "time", "endTime" slots}
  \item{id}{an artificial Iid}
}}
\description{
A dataset containing the track of a z bird.
}
\keyword{data}
