% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoscan.R
\name{stoscan}
\alias{stoscan}
\title{Spatio-temporal clustering from tdbscan outpout manipulated by dt2Convexhull}
\usage{
stoscan(
  d,
  pid = "pid",
  arrival = "arrival",
  departure = "departure",
  geometry = "geometry"
)
}
\arguments{
\item{d}{a data.frame including points which belong to clusters (tdbscan output), latitude, longitude and datetime}

\item{pid}{ID unique for points that belong to one cluster}

\item{arrival}{arrival datetime in polygon}

\item{departure}{departure time from polygon}

\item{geometry}{convex hull of points of a unique cluster}
}
\value{
data.table including pid, s_clustID = ID unique for spatial overlapping clusters,
        st_clustID = ID unique for spatial temporal overlapping clusters
}
\description{
Spatio-temporal clustering from tdbscan outpout manipulated by dt2Convexhull
}
\examples{
require(data.table)
require(ggplot2)

data(zbird)
z = tdbscan(zbird, eps = 12, minPts   = 5, maxLag = 5, borderPoints = TRUE )
z = z[, clustID := factor(clustID)]

d = data.frame(zbird) \%>\% data.table
d = merge(z, d, by.x = 'id', by.y = 'sp.ID')

d = rbindlist(list(copy(d[, tagID := 'bird1']), copy(d[, tagID := 'bird2'])), use.names = TRUE)
d[tagID == 'bird2', x := x + 5]
d[tagID == 'bird2', x := x + 5]
d[!is.na(clustID), ID := paste0(tagID, '_', clustID)]

dp = dt2Convexhull(d, pid = 'ID', projection = '+proj=utm +zone=4 +datum=WGS84')

s = stoscan(dp)
d = merge(d, s, by.x = 'ID', by.y = 'pid', all.x = TRUE)

# plot of spatial overlapping clusters
ggplot(d, aes(x, y, color = as.character(s_clustID) ) ) +
  geom_path(aes(color = NULL), col = 'grey', size = .5) +
  geom_point( alpha = .5, size = 2)

# plot of spatial and temporal overlapping clusters
ggplot(d, aes(x, y, color = as.character(st_clustID) ) ) +
  geom_path(aes(color = NULL), col = 'grey', size = .5) +
  geom_point( alpha = .5, size = 2)

}
